+++ openquake/ghm/create_map_from_curves.py
@@ -31,7 +31,7 @@ import glob
 import numpy as np
 import geopandas as gpd
 from openquake.baselib import sap
-from openquake.man.useful_scripts import csv_output_utils as csvt
+from openquake.man.utilities import csv_output_utils as csvt
 
 
 def read_hazard_curve_files(path_in, prefix=''):
